Fix insight opt-in prompt to use prompt 0.1
Previously the insight prompt has been trying to use prompt 0.2, while the
rest of the project wants to use prompt 0.1 (to match the version required by
grunt).

As a result, insight.js was getting disappointed, and expressed that by
displaying 'undefined' instead of the opt-in prompt. This commit makes the
insight prompt happy again.