setup/install.sh: don't assume Mac if we're running on Linux
If a package manager couldn't be found then the install script would
silently assume the user is running on a Mac. The 'uname' program should
be available on all Linux platforms and can be used to determine the
kernel. If we can't find a package manage, uname is present and it
claims we're running on Linux, print an apology and give up.