Load "test" static files aswell
I need to load static files from "test" directory to run e2e tests on the browser. Runner.html is the file a need to load to launch my e2e tests. This file is under "test" directory (which is at the same level of my "app" directory) and has to be here. I need to load "test" static files in my "app" for two reasons. 1 - This way I'm able to navigate to runner.html even if runner.html is not in the "app" directory. 2 - In my tests I need to navigate to paths relatives to "app" and this way all paths are correct.